Remove intermediate Vec used for branches
The vec previously returned from `Alignments::branches` was not necessary. It
served only to pass values between it and the calling fn `Alignments::next`.
Removing this vec and changing the fn into `add_branches_to_stack`, making it
push items directly onto the stack of the `Alignments` provides a ~10% boost in
performance due to reduced allocation pressure.